backend/requirements.txt
# Core dependencies for RAG content ingestion pipeline
# Generated with: uv export --no-header --no-emit-project

beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
cohere==5.0.2
idna==3.6
requests==2.31.0
qdrant-client>=1.7.0
python-dotenv==1.0.0
tiktoken==0.7.0
urllib3==2.1.0

# Feature 007: RAG Agent Backend dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
google-generativeai==0.3.2
asyncpg==0.29.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: Development dependencies (not needed for production)
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.11.0
# ruff==0.1.11
# mypy==1.7.0
